# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PawZone API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get orders for current user (role-based)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [buyer, seller, transporter]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order from cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderBody"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/payment:
    post:
      operationId: processPayment
      tags: [orders]
      summary: Simulate payment for an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentBody"
      responses:
        "200":
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/issue:
    post:
      operationId: reportIssue
      tags: [orders]
      summary: Report an issue with an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportIssueBody"
      responses:
        "201":
          description: Issue reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review for a seller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/seller/{sellerId}:
    get:
      operationId: getSellerReviews
      tags: [reviews]
      summary: Get reviews for a seller
      parameters:
        - name: sellerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        buyerId:
          type: integer
        buyerName:
          type: string
        sellerId:
          type: integer
        sellerName:
          type: string
        transporterId:
          type: integer
          nullable: true
        transporterName:
          type: string
          nullable: true
        transporterPhone:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, ready, picked_up, in_transit, delivered, cancelled, refunded]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        subtotal:
          type: number
        platformFee:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        deliveryAddress:
          type: string
        petCode:
          type: string
          nullable: true
        barcodeUrl:
          type: string
          nullable: true
        pickupTime:
          type: string
          format: date-time
          nullable: true
        deliveryTime:
          type: string
          format: date-time
          nullable: true
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        paymentDeadline:
          type: string
          format: date-time
          nullable: true
        inventoryLockedUntil:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderNumber
        - buyerId
        - buyerName
        - sellerId
        - sellerName
        - status
        - paymentStatus
        - subtotal
        - platformFee
        - deliveryFee
        - total
        - deliveryAddress
        - createdAt
    OrderDetail:
      allOf:
        - $ref: "#/components/schemas/Order"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderItem"
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/OrderTimelineEvent"
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
        listingTitle:
          type: string
        category:
          type: string
        breed:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        photos:
          type: array
          items:
            type: string
      required:
        - id
        - listingId
        - listingTitle
        - category
        - breed
        - quantity
        - unitPrice
        - subtotal
        - photos
    OrderTimelineEvent:
      type: object
      properties:
        status:
          type: string
        timestamp:
          type: string
          format: date-time
        note:
          type: string
          nullable: true
      required:
        - status
        - timestamp
    PlaceOrderBody:
      type: object
      properties:
        deliveryAddressId:
          type: integer
        deliveryAddress:
          type: string
      required:
        - deliveryAddress
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
        note:
          type: string
      required:
        - status
    PaymentBody:
      type: object
      properties:
        method:
          type: string
          enum: [upi, card, netbanking]
        amount:
          type: number
      required:
        - method
        - amount
    ReportIssueBody:
      type: object
      properties:
        issueType:
          type: string
        description:
          type: string
      required:
        - issueType
        - description
...
    Review:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        buyerId:
          type: integer
        buyerName:
          type: string
        sellerId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - buyerId
        - buyerName
        - sellerId
        - rating
        - createdAt
    CreateReviewBody:
      type: object
      properties:
        orderId:
          type: integer
        sellerId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
      required:
        - orderId
        - sellerId
        - rating
...